organized unit tests to match main packages
