+++ src/test/java/module-info.java
@@ -6,12 +6,13 @@ module fastj.library.test {
 
     opens unittest.testcases.graphics to org.junit.platform.commons;
     opens unittest.testcases.graphics.game to org.junit.platform.commons;
+    opens unittest.testcases.graphics.gradients to org.junit.platform.commons;
     opens unittest.testcases.graphics.util to org.junit.platform.commons;
-    opens unittest.testcases.graphics.util.gradients to org.junit.platform.commons;
+
+    opens unittest.testcases.input.keyboard to org.junit.platform.commons;
 
     opens unittest.testcases.math to org.junit.platform.commons;
 
     opens unittest.testcases.systems.audio to org.junit.platform.commons;
     opens unittest.testcases.systems.control to org.junit.platform.commons;
-    opens unittest.testcases.systems.input.keyboard to org.junit.platform.commons;
 }
